Rename metamacro.h filename with SDmetamacro.h
