+++ SDWebImage/Private/SDmetamacros.h
